Use `u32` in `MenuEntryActivated` event

+++ src/platform_impl/windows/event_loop.rs
@@ -861,7 +861,7 @@ unsafe fn public_window_callback_inner<T: 'static>(
 
         winuser::WM_COMMAND => {
             use crate::event::WindowEvent::MenuEntryActivated;
-            let id = LOWORD(wparam as DWORD) as isize;
+            let id = LOWORD(wparam as DWORD) as u32;
 
             subclass_input.send_event(Event::WindowEvent {
                 window_id: RootWindowId(WindowId(window)),
